+++ src/allotropy/exceptions.py
