+++ src/allotropy/parser_factory.py
@@ -2,7 +2,7 @@ from datetime import tzinfo
 from enum import Enum
 from typing import Optional, Union
 
-from allotropy.allotrope.exceptions import AllotropeConversionError
+from allotropy.exceptions import AllotropeConversionError
 from allotropy.parsers.agilent_gen5.agilent_gen5_parser import AgilentGen5Parser
 from allotropy.parsers.appbio_absolute_q.appbio_absolute_q_parser import (
     AppbioAbsoluteQParser,
