+++ src/allotropy/parsers/appbio_quantstudio/views.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from collections.abc import Iterator
 from typing import Generic, Optional, TypeVar, Union
 
-from allotropy.allotrope.exceptions import AllotropeConversionError
+from allotropy.exceptions import AllotropeConversionError
 
 T = TypeVar("T")
