+++ tests/parsers/moldev_softmax_pro/to_allotrope_test.py
@@ -2,7 +2,7 @@ import re
 
 import pytest
 
-from allotropy.allotrope.exceptions import AllotropeConversionError
+from allotropy.exceptions import AllotropeConversionError
 from allotropy.parser_factory import Vendor
 from tests.parsers.test_utils import from_file, validate_contents, validate_schema
